DOC-216. Added basic components for login, register, dir list
