+++ frontend/datasafe-ui/src/app/app.module.ts
@@ -1,0 +1,35 @@
+import './polyfills';
+
+import {HttpClientModule} from '@angular/common/http';
+import {NgModule} from '@angular/core';
+import {FormsModule, ReactiveFormsModule} from '@angular/forms';
+import {MatNativeDateModule} from '@angular/material/core';
+import {BrowserModule} from '@angular/platform-browser';
+import {platformBrowserDynamic} from '@angular/platform-browser-dynamic';
+import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
+import {DemoMaterialModule} from './material-module';
+
+import {TreeDynamicExample} from './tree-dynamic-example';
+import { UserComponent } from './user/user.component';
+import { LoginComponent } from './login/login.component';
+import { DirComponent } from './dir/dir.component';
+import { RegisterComponent } from './register/register.component';
+
+@NgModule({
+  imports: [
+    BrowserModule,
+    BrowserAnimationsModule,
+    FormsModule,
+    HttpClientModule,
+    DemoMaterialModule,
+    MatNativeDateModule,
+    ReactiveFormsModule,
+  ],
+  entryComponents: [LoginComponent],
+  declarations: [TreeDynamicExample, UserComponent, LoginComponent, DirComponent, RegisterComponent],
+  bootstrap: [LoginComponent],
+  providers: []
+})
+export class AppModule {}
+
+platformBrowserDynamic().bootstrapModule(AppModule);
