+++ frontend/datasafe-ui/src/index.html
@@ -1,17 +1,11 @@
 <html>
-  <head>
+<head>
     <title>Angular App</title>
-  </head>
-  <body>
-  <link href="https://fonts.googleapis.com/icon?family=Material+Icons&display=block" rel="stylesheet">
+</head>
+<body>
+<link href="https://fonts.googleapis.com/icon?family=Material+Icons&display=block" rel="stylesheet">
 <div class="mat-app-background basic-container">
-  <tree-dynamic-example>loading</tree-dynamic-example>
+    <app-login>loading</app-login>
 </div>
-<span class="version-info">Current build: 8.0.2</span>
-
-
-<!-- Copyright 2019 Google Inc. All Rights Reserved.
-    Use of this source code is governed by an MIT-style license that
-    can be found in the LICENSE file at http://angular.io/license -->
-  </body>
+</body>
 </html>
