--- frontend/datasafe-ui/src/polyfills.ts
@@ -1,10 +1,0 @@
-import 'core-js/es6/reflect';
-import 'core-js/es7/reflect';
-import 'zone.js/dist/zone';
-import 'hammerjs';
-import 'web-animations-js';
-
-
-/**  Copyright 2019 Google Inc. All Rights Reserved.
-    Use of this source code is governed by an MIT-style license that
-    can be found in the LICENSE file at http://angular.io/license */
